pyproject.toml
[tool.poetry]
authors = ["admin <admin@example.com>"]
description = "FastAPI project generated using minimal-fastapi-postgres-template."
name = "app"
version = "0.1.0-alpha"

[tool.poetry.dependencies]
python = "^3.13"

alembic = "^1.14.1"
asyncpg = "^0.30.0"
bcrypt = "^4.3.0"
fastapi = "^0.115.10"
pydantic = { extras = ["dotenv", "email"], version = "^2.10.6" }
pydantic-settings = "^2.8.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
langchain-chroma = "^0.2.3"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.12"
freezegun = "^1.5.1"
greenlet = "^3.1.1"
httpx = "^0.28.1"
mypy = "^1.15.0"
pre-commit = "^4.1.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
pytest-xdist = "^3.6.1"
ruff = "^0.9.9"
uvicorn = { extras = ["standard"], version = "^0.34.0" }

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "-vv -n auto --cov --cov-report xml --cov-report term-missing --cov-fail-under=100"
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.coverage.run]
concurrency = ["greenlet"]
omit = ["app/tests/*"]
source = ["app"]

[tool.mypy]
python_version = "3.13"
strict = true

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
# pycodestyle, pyflakes, isort, pylint, pyupgrade
ignore = ["E501"]
select = ["E", "F", "I", "PL", "UP", "W"]
